+++ tests/covalent_tests/workflow/electron_test.py
@@ -24,6 +24,7 @@ import covalent as ct
 from covalent._shared_files.context_managers import active_lattice_manager
 from covalent._workflow.electron import Electron
 from covalent._workflow.transport import TransportableObject, _TransportGraph
+from covalent.executor.executor_plugins.local import LocalExecutor
 
 
 @ct.electron
@@ -117,3 +118,30 @@ def test_injected_inputs_are_not_in_tg():
 
     assert list(g.nodes) == [0, 1]
     assert list(g.edges) == [(1, 0, 0)]
+
+
+def test_metadata_in_electron_list():
+    """Test if metadata of the created electron list apart from executor is set to default values"""
+
+    def identity(y):
+        return y
+
+    calldep = ct.DepsCall(identity, args=[5], retval_keyword="y")
+
+    @ct.electron(call_before=[calldep], executor=LocalExecutor())
+    def task(x, y=0):
+        return (x, y)
+
+    @ct.lattice
+    def workflow(x):
+        return task(x)
+
+    workflow.build_graph([2])
+
+    task_metadata = workflow.transport_graph.get_node_value(0, "metadata")
+    e_list_metadata = workflow.transport_graph.get_node_value(1, "metadata")
+
+    assert list(e_list_metadata["call_before"]) == []
+    assert list(e_list_metadata["call_after"]) == []
+
+    assert e_list_metadata["executor"] == task_metadata["executor"]